requirements.txt
absl-py
aioredis==1.3.1
amqp==2.6.1
arabic-reshaper==2.1.1
asgi-redis==1.1.0
asgiref==3.3.1
astroid==2.4.2
astunparse==1.6.3
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.7
beautifulsoup4==4.9.3
billiard
boto3==1.16.28
botocore==1.19.43
braintree==4.5.0
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.2.0
celery
certifi==2020.12.5
cffi==1.14.4
channels==3.0.0
channels-redis==3.2.0
chardet==3.0.4
click
cmake==3.20.2
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography
daphne==3.0.1
DateTimeRange==1.0.0
decorator==5.0.9
Django==3.1.4
django-filter==2.4.0
django-model-utils==4.1.1
django-pdfkit==0.3.1
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
django-ses==1.0.3
django-storages==1.10.1
djangorestframework==3.12.2
django-admin-interface==0.16.3
dlib==19.22.0
drf-writable-nested==0.6.2
dynaconf==3.1.4
fcm-django==0.3.7
firebase-admin==4.5.0
Flask==1.1.2
Flask-SocketIO==4.3.1
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.4
google-cloud-core==1.4.4
google-cloud-firestore==2.0.2
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
greenlet
grpcio
gunicorn
h5py
hiredis==1.1.0
html5lib==1.1
httplib2==0.18.1
hvac==0.10.12
hyperlink==20.0.1
idna==2.10
importlib-metadata==3.3.0
imutils==0.5.4
incremental==17.5.0
isort==5.6.4
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
Keras-Preprocessing==1.1.2
knox==0.1.11
kombu
lazy-object-proxy==1.4.3
linkedin-scraper==2.8.2
loguru==0.5.3
lxml==4.6.3
Markdown==3.3.3
MarkupSafe==1.1.1
mbstrdecoder==1.0.1
mccabe==0.6.1
msgpack==1.0.0
msgpack-python==0.5.6
numpy
oauthlib==3.1.0
openapi-codec==1.3.2
opencv-contrib-python
opencv-python
opt-einsum==3.3.0
pandas
pathlib-mate==1.0.1
pdfkit==0.6.0
Pillow==8.0.1
praw==7.2.0
prawcore==2.0.0
proto-plus==1.13.0
protobuf==3.14.0
psycopg2-binary
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyfcm==1.4.9
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL==20.0.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2021.1
qrcode==6.1
redis
reportlab==3.5.56
requests
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
selenium==3.141.0
service-identity==18.1.0
shortuuid==1.0.1
simplejson==3.17.2
six==1.15.0
slacker==0.9.65
slacker-log-handler==1.8.0
soupsieve==2.2
SQLAlchemy==1.4.0
sqlparse==0.4.1
stripe==2.55.1
tensorboard==2.9.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.9.1
tensorflow-estimator==2.9.0rc0
termcolor==1.1.0
toml==0.10.2
tqdm==4.60.0
twilio==6.50.0
Twisted==20.3.0
txaio==20.4.1
typed-ast
typepy==1.1.2
typing-extensions==3.7.4.3
update-checker==0.18.0
uritemplate==3.0.1
urllib3==1.26.2
uszipcode==0.2.5
validators==0.18.2
vine==1.3.0
webencodings==0.5.1
websocket-client==1.0.0
Werkzeug==1.0.1
wrapt==1.12.1
xhtml2pdf==0.2.5
zipp==3.4.0
zope.interface==5.2.0
django-cors-headers
webdrivermanager==0.10.0
webdriver-manager==3.5.2
# gitlab
pyathena==2.13.0
djangorestframework-simplejwt==5.2.0
